Split record_function_bridge installation from processing
