+++ projects/rocprofiler-compute/src/lib/torch_trace_collector/capture_buffer.h
@@ -66,4 +66,8 @@ private:
 
 inline CaptureBuffer g_capture;
 
+// Start and stop recording of the emitted wire strings.
+void                     start_capture();
+std::vector<std::string> stop_capture();
+
 }  // namespace torch_trace_collector::detail
